+++ Frends.ODBC.ExecuteQuery/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [3.0.0] - 2026-02-18
 ### Added
-- Added DataReader wrapper class to avoid assembly reference issues when using DataReader output mode.
+- Added DataReaderWrapper class to encapsulate OdbcDataReader and avoid assembly reference issues when using DataReader output mode.
+- Result now implements IAsyncDisposable and exposes a DisposeAsync method for async resource cleanup.
+
 ### Changed
-- Updated target framework from net6.0 to net8.0.
+- [Breaking change] Updated target framework from net6.0 to net8.0.
 
 ## [2.1.0] - 2025-03-25
 ### Changed